+++ app/src/main/java/com/internetEnemies/combatCritters/Logic/market/MarketCycle.java
@@ -1,4 +1,4 @@
-package com.internetEnemies.combatCritters.Logic;
+package com.internetEnemies.combatCritters.Logic.market;
 
 
 import com.internetEnemies.combatCritters.data.IMarketDB;
