+++ app/src/test/java/com/internetEnemies/combatCritters/IntegrationTests/MarketCycleIntegrationTest.java
@@ -4,10 +4,10 @@ import static junit.framework.TestCase.assertEquals;
 
 import static org.junit.Assert.assertNotEquals;
 
-import com.internetEnemies.combatCritters.Logic.IMarketCycle;
-import com.internetEnemies.combatCritters.Logic.IMarketHandler;
-import com.internetEnemies.combatCritters.Logic.MarketCycle;
-import com.internetEnemies.combatCritters.Logic.MarketHandler;
+import com.internetEnemies.combatCritters.Logic.market.IMarketCycle;
+import com.internetEnemies.combatCritters.Logic.market.IMarketHandler;
+import com.internetEnemies.combatCritters.Logic.market.MarketCycle;
+import com.internetEnemies.combatCritters.Logic.market.MarketHandler;
 import com.internetEnemies.combatCritters.Logic.transaction.TransactionHandler;
 import com.internetEnemies.combatCritters.TestUtils;
 import com.internetEnemies.combatCritters.data.IMarketDB;
